backend/requirements.txt
absl-py                 
albucore                
albumentations          
annotated-types         
anyio                   
appnope                 
asttokens               
astunparse              
blinker                 
certifi                 
charset-normalizer      
click                   
comm                    
contourpy               
cycler                  
debugpy                 
decorator              
distro                 
easyocr                 
exceptiongroup         
executing               
filelock                
Flask                   
flask-cors              
flatbuffers             
fonttools               
fsspec                  
gast                    
google-pasta            
grpcio                  
h11                     
h5py                    
httpcore                
httpx                   
idna                    
imageio                
ipykernel               
ipython                 
itsdangerous            
jedi                    
Jinja2                 
jiter                   
joblib                  
jupyter_client          
jupyter_core            
keras                   
kiwisolver              
lazy_loader             
libclang                
Markdown                
markdown-it-py          
MarkupSafe             
matplotlib              
matplotlib-inline       
mdurl                  
ml_dtypes               
mpmath                  
namex                  
nest-asyncio            
networkx                
ninja                   
numpy                  
openai                  
opencv-python           
opencv-python-headless  
opt_einsum              
optree                 
packaging              
pandas                  
parso                   
pexpect                 
pillow                  
pip                     
platformdirs            
prompt_toolkit          
protobuf                
psutil                  
ptyprocess              
pure_eval              
pyclipper               
pydantic                
pydantic_core           
Pygments                
pyparsing               
pytesseract             
python-bidi             
python-dateutil         
pytz                    
PyYAML                 
pyzmq                   
requests                
rich                    
scikit-image            
scikit-learn            
scipy                   
seaborn                 
setuptools              
shapely                 
simsimd                 
six                     
sniffio                 
stack-data              
stringzilla             
sympy                   
tensorboard             
tensorboard-data-server 
tensorflow              
termcolor               
threadpoolctl           
tifffile                
torch                   
torchvision             
tornado                 
tqdm                    
traitlets               
typing_extensions       
tzdata                  
urllib3                 
wcwidth                 
Werkzeug                
wheel                  
wrapt                 